closes #12

+++ src/executor/count.rs
@@ -10,7 +10,7 @@ use crate::store::kv::{StoreKVPool, STORE_ACCESS_LOCK};
 pub struct ExecutorCount;
 
 impl ExecutorCount {
-    pub fn execute<'a>(store: StoreItem<'a>) -> Result<u64, ()> {
+    pub fn execute<'a>(store: StoreItem<'a>) -> Result<u32, ()> {
         if let StoreItem(collection, _bucket_value, _object_value) = store {
             // Important: acquire database access read lock, and reference it in context. This \
             //   prevents the database from being erased while using it in this block.
