+++ src/executor/flushc.rs
@@ -10,7 +10,7 @@ use crate::store::kv::StoreKVActionBuilder;
 pub struct ExecutorFlushC;
 
 impl ExecutorFlushC {
-    pub fn execute<'a>(store: StoreItem<'a>) -> Result<u64, ()> {
+    pub fn execute<'a>(store: StoreItem<'a>) -> Result<u32, ()> {
         // Important: do not acquire the store from there, as otherwise it will remain open \
         //   even if dropped in the inner function, as this caller would still own a reference to \
         //   it.
